feat: implement analysis interfaces and update signal handling for historical data
